NutriGraph/prototypes/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
networkx>=3.0
pyvis>=0.3.1
matplotlib>=3.7.0
langchain-google-genai
pillow
